refactor: Rename test files and implement uniform matrix operations in new modules
